// src/app/core/interfaces/series.interface.ts
export interface ISeries {
    id: number;
    title: string;
    description: string;
    thumbnail: {
        path: string;
        extension: string;
    };
    pageCount: number;
    creators: {
        available: number;
        collectionURI: string;
        items: {
            resourceURI: string;
            name: string;
            role: string;
        }[];
    };
    dates: {
        type: string;
        date: string;
    }[];
}